requirements.txt
numpy
Flask
gunicorn
sklearn
pandas
ipykernel
PyYAML